requirements.txt
requests~=2.32.4
uuid~=1.30
boto3~=1.38.38
kafka-python~=2.2.15
botocore~=1.38.46